pyproject.toml
[tool.poetry]
name = "zhenxun-utils"
version = "0.1.0"
description = "小真寻bot的工具类"
authors = ["HibiKier <775757368@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
strenum = "^0.4.15"
pillow = "^9.0.0"
httpx = "^0.27.2"
rich = "^13.9.4"
aiofiles = "^24.1.0"
nonebot-plugin-htmlrender = "^0.4.0"
nonebot-plugin-alconna = "^0.53.1"
retrying = "^1.3.4"
nonebot-plugin-uninfo = "^0.5.0"
pypinyin = "^0.53.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
